+++ NEWS.md
@@ -1,3 +1,8 @@
+# ROMOPAPI 2.1.0
+- Added getAPIInfo endpoint to API
+- Updated db with large tree
+- Fixed error in getCodeCounts when conceptId was the root of the tree
+
 # ROMOPAPI 2.0.0
 - Updated to use DatabaseConnector v7